Extract test util `choose_reply` to `tests.api.mocks`
