+++ tests/api/mocks.py
@@ -12,15 +12,62 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import collections
 import io
 import uuid
 from typing import Any, Optional, Tuple, Union
+from unittest import mock
+
+import orjson
+from requests.exceptions import HTTPError
+from requests.structures import CaseInsensitiveDict
 
 from dwave.cloud.coders import encode_problem_as_qp, encode_problem_as_ref
 from dwave.cloud.utils.qubo import generate_const_ising_problem
 from dwave.cloud.utils.time import utcrel
 from dwave.cloud.solver import StructuredSolver, UnstructuredSolver
 from dwave.cloud.testing.mocks import qpu_clique_solver_data, hybrid_bqm_solver_data
+
+
+def choose_reply(path, replies, statuses=None, date=None):
+    """Choose the right response based on the path and make a mock response."""
+
+    if statuses is None:
+        statuses = collections.defaultdict(lambda: iter([200]))
+
+    if date is None:
+        date = utcrel(0)
+
+    if path in replies:
+        response = mock.Mock(['content', 'text', 'json', 'raise_for_status', 'headers'])
+        response.status_code = next(statuses[path])
+        content = replies[path]
+        if isinstance(content, str):
+            content = content.encode('utf8')
+        elif not isinstance(content, bytes):
+            content = orjson.dumps(content)
+        response.content = content
+        response.text = content.decode('utf8')
+        response.json.side_effect = lambda: replies[path]
+        response.headers = CaseInsensitiveDict({'Date': date.isoformat()})
+
+        def raise_for_status():
+            if not 200 <= response.status_code < 400:
+                raise HTTPError(response.status_code)
+        response.raise_for_status = raise_for_status
+
+        def ok():
+            try:
+                response.raise_for_status()
+            except HTTPError:
+                return False
+            return True
+        ok_property = mock.PropertyMock(side_effect=ok)
+        type(response).ok = ok_property
+
+        return response
+    else:
+        raise NotImplementedError(path)
 
 
 class SapiMockResponses:
