+++ tests/test_upload.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import collections
 import io
 import json
 import os
@@ -22,9 +21,7 @@ import unittest
 from concurrent.futures import ThreadPoolExecutor, wait
 from unittest import mock
 
-import orjson
 from parameterized import parameterized
-from requests.exceptions import HTTPError
 
 from dwave.cloud.client import Client
 from dwave.cloud.exceptions import SAPIError, ProblemUploadError
@@ -33,6 +30,7 @@ from dwave.cloud.upload import (
 from dwave.cloud.utils.time import tictoc
 
 from tests import config
+from tests.api.mocks import choose_reply
 
 
 class TestGettableABC(unittest.TestCase):
@@ -410,47 +408,6 @@ class TestChunkedData(unittest.TestCase):
         chunks_expected = [b'012', b'345', b'678', b'9']
         chunks_generated = [g().read() for g in cd.generators()]
         self.assertListEqual(chunks_expected, chunks_generated)
-
-
-def choose_reply(path, replies, statuses=None):
-    """Choose the right response based on a hashable `key` and make a mock
-    response.
-    """
-    # dedup vs test_mock_submissions.choose_reply
-
-    if statuses is None:
-        statuses = collections.defaultdict(lambda: iter([200]))
-
-    if path in replies:
-        response = mock.Mock(['content', 'text', 'json', 'raise_for_status', 'headers'])
-        response.status_code = next(statuses[path])
-        content = replies[path]
-        if isinstance(content, str):
-            content = content.encode('utf8')
-        elif not isinstance(content, bytes):
-            content = orjson.dumps(content)
-        response.content = content
-        response.text = content.decode('utf8')
-        response.json.side_effect = lambda: replies[path]
-        response.headers = {}
-
-        def raise_for_status():
-            if not 200 <= response.status_code < 400:
-                raise HTTPError(response.status_code)
-        response.raise_for_status = raise_for_status
-
-        def ok():
-            try:
-                response.raise_for_status()
-            except HTTPError:
-                return False
-            return True
-        ok_property = mock.PropertyMock(side_effect=ok)
-        type(response).ok = ok_property
-
-        return response
-    else:
-        raise NotImplementedError(path)
 
 
 @mock.patch('time.sleep', lambda *args: None)
